Fixed UE on FormCreditRecurringCharges when payment plan amounts were not whole numbers. Backported to 15.3 Reviewed by: Ryan, Cameron
git-svn-id: https://opendentalsoft.com:23793/svn/opendental/opendental15.3@12673 098d0599-7f2b-0410-95a5-ad30ac85393a

+++ OpenDental/Forms/FormCreditRecurringCharges.cs
@@ -100,11 +100,11 @@ namespace OpenDental {
 					}
 					if(table.Rows[i]["PayOrder"].ToString()=="2") {
 						famBalTotal+=rptChargeAmt;
-						rptChargeAmt+=PIn.Long(table.Rows[i]["ChargeAmt"].ToString());
+						rptChargeAmt+=PIn.Double(table.Rows[i]["ChargeAmt"].ToString());
 					}
 				}
 				else {
-					rptChargeAmt=PIn.Long(table.Rows[i]["ChargeAmt"].ToString());
+					rptChargeAmt=PIn.Double(table.Rows[i]["ChargeAmt"].ToString());
 				}
 				if(!dictPatBals.ContainsKey(patNum)) {
 					dictPatBals.Add(patNum,famBalTotal);
